Creacion de objetos y enemigos
Primera parte de creacion de objetos y enemigos en escenario

+++ WEB2023/Assets/Scripts/DungeonGenerator/Room.cs
@@ -6,8 +6,12 @@ public class Room : MonoBehaviour
 {
 
 
-    public float Width;
-    public float Height;
+    //public float Width;
+    public int Width;
+
+
+    //public float Height;
+    public int Height;
 
     public int X;
     public int Y;
